Add player drawing

+++ src/math/radians.c
@@ -2,7 +2,7 @@
 #include "../../headers/cubed.h"
 #include <math.h>
 
-double	ft_dtorad(int angle)
+double	ft_dtorad(double angle)
 {
-	return (angle * M_PI/180.0);
+	return (angle * M_PI / 180.0);
 }
